+++ debug.py
@@ -1,4 +1,4 @@
-from rule34Py import rule34Py, version
+from rule34Py import rule34Py
 import json
 #import random
 
@@ -13,11 +13,12 @@ result_icame = r34Py.icame()
 result_search = r34Py.search(["neko"], page_id=2, limit=50)
 result_pool = r34Py.get_pool(17509) # or r34Py.get_pool(17509, false)
 result_random = r34Py.random_post(["neko"]) # or r34Py.random_post()
-result_tagmap = r34Py.tagmap()
+result_tag_map = r34Py.tag_map()
+result_top_tags = r34Py.top_tags()
 
 print(result_random.id)
 print(result_random.image)
 
-print(result_icame[0].character_name) # returns the character name of the first item
+print(result_icame[0].character_name, result_icame[0].count) # returns the character name of the first item
 
-print(result_tagmap[0].tagname)
+print(result_top_tags[0].tagname)